refactor: add champion fns to champion.c
