+++ include/champion.h
@@ -8,6 +8,7 @@
 #include "op.h"
 #include "instructions.h"
 #include "vm.h"
+#include "utils.h"
 
 champion_t *init_champion();
 
@@ -18,5 +19,9 @@ void add_champion(core_t *core_t, champion_t *champion);
 champion_t* clone_champion(const champion_t* original);
 
 int add_champion_to_core(core_t* core, champion_t* champion);
+
+champion_t *find_champion(core_t *core_vm, int id);
+
+int check_champion_lives(core_t *core_vm);
 
 #endif
